Implement Dockerfile for Team Microservice

+++ Trainer_microservice/src/main/java/com/icesi/edu/co/repository/TrainerRepository.java
@@ -3,6 +3,6 @@ package com.icesi.edu.co.repository;
 import com.icesi.edu.co.entity.Trainer;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TrainerRepository extends JpaRepository<Trainer, Integer>  {
+public interface TrainerRepository extends JpaRepository<Trainer, Long>  {
     
 }
